[Simulator] MiniBrowser fails to launch with micro version difference from SDK version. https://bugs.webkit.org/show_bug.cgi?id=301600 rdar://163599114
Reviewed by Basuke Suzuki.

When MiniBrowser launches, it checks for SDK and runtime version being equal.
This means that differences in micro version between each will cause the check
to fail.

This change modifies this subroutine to check if the runtime version starts
with the SDK version if they're not strictly equal. If so, we warn the user
that we're fuzzy matching the runtime.

* Tools/Scripts/webkitdirs.pm:
    (simulatorRuntime): Fuzzy match if runtime version starts with SDK version if inequal.

Canonical link: https://commits.webkit.org/302281@main