[GTK] Enable MSAA for GPU rendering https://bugs.webkit.org/show_bug.cgi?id=301273
Reviewed by Miguel Gomez

Only keep it disabled for old intel chips where MSAA performance is not
good enough. Also disable dynamic MSAA for other ports, since results
in embedded are not that good, but adding an env var so that it can be
enabled or disabled in any port.

Canonical link: https://commits.webkit.org/302289@main