[WebXR][GLib] TestWebKitWebXR/permission-request is flaky https://bugs.webkit.org/show_bug.cgi?id=301508
Reviewed by Adrian Perez de Castro.

TestWebKitWebXR/permission-request test was flaky on the Buildbot, particularly
on EWS. The result window title was "" rather than the expected strings "pass"
or "fail". This was reproducible on my PC by limiting the CPU affinity with
taskset.
> while WITH_OPENXR_RUNTIME=y taskset 1 ./Tools/Scripts/run-gtk-tests --release WebKitGTK/TestWebKitWebXR; do; done

In the test code, the title was cleared after the test with the following code:
> test->runJavaScriptAndWaitUntilFinished("document.title = ''", nullptr);

The problem was that we didn't wait the title changed. We should do:
> test->waitUntilTitleChanged();

Rather than changing the title, this change loads an empty string before the
test run.

* Tools/TestWebKitAPI/Tests/WebKitGLib/TestWebKitWebXR.cpp:
(testWebKitXRPermissionRequest):
* Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/302292@main