Fix Webkit.org survey plugin to show survey and results https://bugs.webkit.org/show_bug.cgi?id=301635
Reviewed by Devin Rousso.

Fixes HTML in the survey results that prevented rendering the results
correctly. This includes a drive-by fix for the table-of-contents plugin
that prevented the page from saving correctly in some cases.

Additional changes enhance how text responses for surveys are taken to make
the experience more inline with expectations.

* Websites/webkit.org/wp-content/plugins/table-of-contents.php:
* Websites/webkit.org/wp-content/plugins/webkit-survey/plugin.php:
* Websites/webkit.org/wp-content/plugins/webkit-survey/results.php:
* Websites/webkit.org/wp-content/plugins/webkit-survey/survey.php:

Canonical link: https://commits.webkit.org/302298@main