[GLIB] Fix the WebKitWebsiteData/itp test https://bugs.webkit.org/show_bug.cgi?id=301639
Reviewed by Carlos Garcia Campos.

The test was checking that after enabling ITP the database
file and directory don't exist, but enabling ITP sends a
message to the network process that causes the
ResourceLoadStatisticsStore to be instantiated right away,
which does create the directory and database file. The
test might have passed because of the asynchronous nature of
the communication with the network process, but the check
in itself was incorrect.

* Tools/TestWebKitAPI/Tests/WebKitGLib/TestWebsiteData.cpp:
(testWebsiteDataITP):
* Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/302302@main