[WebXR] WPT WebXR/xrWebGLLayer_opaque_framebuffer_stencil incorrectly uses glReadPixels https://bugs.webkit.org/show_bug.cgi?id=301409 rdar://163318566
Reviewed by Kimmo Kinnunen.

The OpenGL specification says only two combinations of format and type for any
internalformat are accepted in most cases. The first varies depending on the
internalformat of the currently bound rendering surface. For the format and
types used in the test, the combination format RGBA and type UNSIGNED_BYTE is
accepted. The second implementation-chosen format and type for this internalformat
may be determined by calling GetIntegerv with the symbolic constants
IMPLEMENTATION_COLOR_READ_FORMAT and IMPLEMENTATION_COLOR_READ_TYPE, which the
test is not doing.

Changing the format passed to glReadPixels from GL_RGB to GL_RBGA fixes the
test and follows the specification.

Renable and update expectations for xrWebGLLayer_opaque_framebuffer_stencil.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/webxr/xrWebGLLayer_opaque_framebuffer_stencil.https-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/webxr/xrWebGLLayer_opaque_framebuffer_stencil.https.html:
* LayoutTests/platform/gtk/TestExpectations:
* LayoutTests/platform/visionos/TestExpectations:
* LayoutTests/platform/wpe/TestExpectations:

Canonical link: https://commits.webkit.org/302321@main