[cleanup] TextBoxPainter::collectDecoratingBoxesForBackgroundPainting should not rely on absolute distance https://bugs.webkit.org/show_bug.cgi?id=301672#
Reviewed by Antti Koivisto.

This change is in preparation for supporting vertical content when
the parent box is not the decorating box.

Currently we compute the decorating box location by simply using its location
-> 'm_paintOffset.y() + inlineBox->logicalTop() + borderAndPaddingBefore'
but this does not work well with vertical content where the graphics context is rotated.
Instead let's compute the diff between these 2 inline boxes and use it to adjust the location.

* Source/WebCore/rendering/TextBoxPainter.cpp:
(WebCore::TextBoxPainter::collectDecoratingBoxesForBackgroundPainting):

Canonical link: https://commits.webkit.org/302345@main