[threaded-animation-resolution] pass timeline data to `AcceleratedEffect::apply()` https://bugs.webkit.org/show_bug.cgi?id=301683 rdar://163703642
Reviewed by Matt Woodrow.

In 302282@main we added the notion of a duration for timelines in the remote layer tree.
As a result, we now have all the required information to pass the timeline current time
and duration down to `AcceleratedEffect::apply()` and we can remove the FIXME to that
effect.

* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
(WebCore::AcceleratedEffect::AcceleratedEffect):
(WebCore::AcceleratedEffect::apply const):
* Source/WebCore/platform/animation/AcceleratedEffect.h:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteAnimation.cpp:
(WebKit::RemoteAnimation::apply):

Canonical link: https://commits.webkit.org/302361@main