[WebXR][WPT] Stop xrWebGLLayer_opaque_framebuffer_stencil sniffing navigator.platform https://bugs.webkit.org/show_bug.cgi?id=301728 rdar://163752529
Reviewed by Mike Wyrzykowski.

The test claims that "mac has issues with readPixels from the default fbo." and
is searching for substring "Mac" in navigator.platform. WebKit reports
'MacIntel' on all Cocoa platforms, so this check is of little use.

The test was incorrectly assuming that GL_RGB was a valid format for use with
glReadPixels, which has been fixed in  https://commits.webkit.org/302321@main

* LayoutTests/imported/w3c/web-platform-tests/webxr/xrWebGLLayer_opaque_framebuffer_stencil.https.html:
* LayoutTests/platform/visionos/imported/w3c/web-platform-tests/webxr/xrWebGLLayer_opaque_framebuffer_stencil.https-expected.txt: Removed.

Canonical link: https://commits.webkit.org/302367@main