REGRESSION (Safari 26): Live HLS ID3 timed metadata cues mapped with endTime = Infinity cause TextTrack.activeCues to include all past cues indefinitely https://bugs.webkit.org/show_bug.cgi?id=299853 rdar://161687917
Reviewed by Jer Noble.

274146@main changed the way SerializedPlatformDataCueValue objects are compared so instead
of creating the native NSObjects and comparing them, every field in the internal Data struct
is compared directly. The `locale` field is an NSLocale, so values are compared with `isEqual:`,
but because any message to a nil NSObject returns a falsy value,
`SerializedPlatformDataCueValue::Data::operator==` always returns false if the data cue
doesn't have a locale.

track-in-band-hls-metadata-cue-duration.html was updated for this fix.

* LayoutTests/http/tests/media/hls/track-in-band-hls-metadata-cue-duration.html:
* Source/WebCore/platform/cocoa/SerializedPlatformDataCueValue.mm:
(WebCore::SerializedPlatformDataCueValue::Data::operator== const):

Canonical link: https://commits.webkit.org/302395@main