[Build Speed] Reduce includes of CSSGradient.h rdar://163804793 https://bugs.webkit.org/show_bug.cgi?id=301764
Reviewed by Brandon Stewart, Simon Fraser, and Sam Weinig.

Prior to this patch, CSSGradient.h was the 15th most expensive header in a
unified WebCore build. It is included 317 times, at an average CPU cost of 369
ms on this machine, for a total CPU cost of nearly 2m.

However this header was included entirely unnecessarily in two other commonly
included headers. The include statements can be removed and replaced with
absolutely nothing.

After this patch, CSSGradient.h is now the 832nd most expensive header in a
unified WebCore build. It is included 11 times, for a total CPU cost of 2s, a
reduction of 98%.

* Source/WebCore/css/values/shapes/CSSCircleFunction.h:
* Source/WebCore/css/values/shapes/CSSEllipseFunction.h:

Canonical link: https://commits.webkit.org/302407@main