[run-jsc-stress-tests] Improve architecture checks https://bugs.webkit.org/show_bug.cgi?id=301753
Reviewed by Carlos Alberto Lopez Perez.

`run-jsc-stress-tests` can inadvertently select the tests to run based
on the host system, whereas we may be testing a cross-compiled binary
(but have neglected to explicitly specify the architecture). This user
error leads to spurious failures; catch the error early instead.

Functionally, bail if the detected architecture (if any) doesn't match
an explicitly specified architecture. Also refuse to continue if we
can't detect the architecture of the JSC binary and the user hasn't
given us an --architecture either. Ensure we never try to run the tests
without knowing which architecture we're testing for.

Canonical link: https://commits.webkit.org/302469@main