[Win] More robust approach to cleaning build dir https://bugs.webkit.org/show_bug.cgi?id=301834
Reviewed by Yusuke Suzuki.

When extracting the built product, sometimes the worker can't delete
some of the DLLs in the build directory. I can't see any processes
holding the lock on the DLLs, so we're stuck until the container is
restarted.

To unblock the build, we'll move the files we can't delete to a
temporary WebKitTrash location. We'll try and delete the WebKitTrash but
ignore any errors - next time the container restarts those files should
get cleaned up.

Canonical link: https://commits.webkit.org/302473@main