[Build Speed] Reduce includes of StyleBasicShape.h rdar://163812454 https://bugs.webkit.org/show_bug.cgi?id=301767
Reviewed by Sam Weinig.

Before this patch, StyleBasicShape.h was the 4th most expensive header in a
Unified WebCore build. It was included 310 times, at an average cost of 1121
ms, for a total CPU cost of 5m47s.

It is included in 4 headers total, and in only one of those can its full
definitions be replaced with forward declarations. That forward declaration is
complex enough to warrant its own header file, StyleShapeForward.h.

After this patch, StyleBasicShape.h is the 24th most expensive header. It is
included 210 times for a total CPU cost of 1m37s.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/rendering/RenderBox.h:
* Source/WebCore/rendering/TextBoxPainter.h:
* Source/WebCore/rendering/shapes/LayoutShape.h:
* Source/WebCore/style/values/shapes/StyleShapeForward.h: Added.

Canonical link: https://commits.webkit.org/302486@main