Allow to iterate over all CoreAudioSharedUnits rdar://163919516 https://bugs.webkit.org/show_bug.cgi?id=301849
Reviewed by Eric Carlson.

We introduce CoreAudioSharedUnit::forEach to allow suspending/resuming/reconfiguring all units.
This is implemented via a WeakHashSet.

We rename CoreAudioSharedUnit::singleton to CoreAudioSharedUnit::defaultSingleton as we plan to progressively restrict this method to the audio unit that is responsible for audio rendering.
To do so, we plan to split of the functionality of CoreAudioSharedUnit into specific classes.

No change of behavior.

Canonical link: https://commits.webkit.org/302520@main