REGRESSION(300463@main): Sending the fill pattern image to GPUP can cause null-pointer dereference in RemoteResourceCacheProxy::recordNativeImageUse() https://bugs.webkit.org/show_bug.cgi?id=301958 rdar://163447308
Reviewed by Simon Fraser.

Some NativeImages may have backing stores in GPUP only. This happens when copying
a NativeImage from a RemoteImageBuffer. Examples for this scenario is the tileImage
of the fill and the stroke patterns.

After 300463@main and when GPUP crashes, these NativeImages can't be recovered
because their backing stores in GPUP are deleted. The originator RemoteImageBuffers
are also deleted. So these images cannot be drawn until a full layout happens.

For now, add a null check to recordNativeImageUse().

* Source/WebKit/WebProcess/GPU/graphics/RemoteResourceCacheProxy.cpp:
(WebKit::RemoteResourceCacheProxy::recordNativeImageUse):

Canonical link: https://commits.webkit.org/302568@main